% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ILS}
\alias{ILS}
\title{The iterative least square estimator}
\usage{
ILS(X, A_init, B_init, max_iters)
}
\arguments{
\item{X}{a tensor of shape (m, n, t), where m is the number of rows, n is the number of columns, and t is the number of observations}

\item{A_init}{the initial value of A}

\item{B_init}{the initial value of B}

\item{max_iters}{the maximum iterations}
}
\value{
a list containing the estimate of A and B
}
\description{
The iterative least square estimator using Rcpp and RcppArmadillo
}
\examples{
\dontrun{
  ils.est <- ILS(X, A_init, B_init, 200)
}
}
