% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MLE}
\alias{MLE}
\title{The maximum likelihood estimator}
\usage{
MLE(A_init, B_init, sigmar_init, sigmac_init, X, max_iters)
}
\arguments{
\item{A_init}{the initial value of A}

\item{B_init}{the initial value of B}

\item{sigmar_init}{the initial value of sigmar}

\item{sigmac_init}{the initial value of sigmac}

\item{X}{a tensor of shape (m, n, t), where m is the number of rows, n is the number of columns, and t is the number of observations}

\item{max_iters}{the maximum iterations}
}
\value{
a list containing the estimate of A, B, sigmac, sigmar
}
\description{
The maximum likelihood estimator using Rcpp and RcppArmadillo
}
\examples{
\dontrun{
  mle.est <- MLE(A.init, B.init, sigmar.init, sigmac.init, X, 200)
}
}
