% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PROJ}
\alias{PROJ}
\title{The projection estimator}
\usage{
PROJ(X)
}
\arguments{
\item{X}{a tensor of shape (m, n, t), where m is the number of rows, n is the number of columns, and t is the number of observations}
}
\value{
a list containing the estimate of A and B
}
\description{
The projection estimator using Rcpp and RcppArmadillo
}
\examples{
\dontrun{
  proj.est <- PROJ(X)
}
}
