% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsc.R
\name{rsc}
\alias{rsc}
\title{rsc}
\usage{
rsc(x, y, min_p_n, sigma, n, q)
}
\arguments{
\item{x}{the matrix x}

\item{y}{the matrix y}

\item{min_p_n}{the minmum of the dimension p and n}

\item{sigma}{stand error of the model's error term}

\item{n}{n}

\item{q}{the rank of the matrix x}
}
\value{
the norm of the matrix x
}
\description{
get the modifiedthe rank of the model,it's the realizationof the rank selection criterion
}
\examples{
\dontrun{
set.seed(1)
A<-matrix(c(runif(80,-2,2)),nrow=16)
A[14:16,]=0
A[,5]=0
x<-matrix(rnorm(16*16),nrow=16)
e<-matrix(rnorm(16*5,0,1),nrow=16)
sigma=1
y=x\%*\%A+e
p=ncol(x)
n=ncol(y)
m=nrow(x)
min_p_n=min(p,n);
q=16;
k=rsc(x,y,min_p_n,sigma,n,q);
}
}
