\name{StatComp20081-package}
\docType{package}
\title{
  The Algorithms in paper "Autoregressive Models for Matrix-Valued Time Series"
}
\description{
  The Algorithms in paper "Autoregressive Models for Matrix-Valued Time Series", including Projection    method, VAR method, iterative least squares method and maximum likelihood method.
}
\details{
  This package implements the algorithms in paper
  R. Chen, H. Xiao and D. Yang, Autoregressive models for matrix-valued time series.
  Journal of Econometrics (2020), https://doi.org/10.1016/j.jeconom.2020.07.015.,
  including Var, projection method, iterated least squares and maximum likelihood estimate.
}
\author{
Yuzhe Zhang

Yuzhe Zhang <zyz2020@mail.ustc.edu.cn>
}
\references{
  R. Chen, H. Xiao and D. Yang, Autoregressive models for matrix-valued time series.
  Journal of Econometrics (2020), https://doi.org/10.1016/j.jeconom.2020.07.015.
}
\keyword{ package }
