% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VAR}
\alias{VAR}
\title{The vector autoregressive estimate}
\usage{
VAR(X)
}
\arguments{
\item{X}{a tensor of shape (m, n, t), where m is the number of rows, n is the number of columns, and t is the number of observations}
}
\value{
the estimate of phi = kronecker(B, A)
}
\description{
The vector autoregressive estimate using Rcpp and RcppArmadillo
}
\examples{
\dontrun{
  phi.hat <- VAR(X)
}
}
